Implement interfaces for mediaapi storage
